moves connectPeers
